This change modifies goog.dom.safe.openInWindow function to not pass opt_specs and opt_replace to window.open if they are both undefined.
The aim of this change is to work around a bug in Internet Explorer 11 which unlike other browsers opens a new window rather than a new tab when those two parameters are passed to window.open, even if they’re undefined. This is described e.g. in https://stackoverflow.com/a/41489586.
RELNOTES: goog.dom.safe.openInWindow now opens the page in a new tab rather than a new window on IE11
PiperOrigin-RevId: 340720646